Thread Pool
I've been programing for 3 weeks now... I wanted to make the software faster, so i wrote it all over again, but now using concurrent futures to create a threading pool. In order to do that i had to put the functions out of classes. i'm not working with objects any more (it seemed so fancy). I'm really proud because it worked! For the fist time i created multiple threads with returning values.